events cleanup 1/2
