+++ packages/fiber/package.json
@@ -40,14 +40,14 @@
     "prebuild": "cp ../../readme.md readme.md"
   },
   "dependencies": {
-    "@babel/runtime": "^7.13.9",
+    "@babel/runtime": "^7.13.10",
     "react-merge-refs": "^1.1.0",
-    "react-reconciler": "^0.26.1",
+    "react-reconciler": "^0.26.2",
     "react-three-fiber": "0.0.0-deprecated",
     "react-use-measure": "^2.0.4",
     "resize-observer-polyfill": "^1.5.1",
     "scheduler": "^0.20.2",
-    "use-asset": "^1.0.3",
+    "use-asset": "^1.0.4",
     "utility-types": "^3.10.0",
     "zustand": "^3.3.3"
   },
